lab3/sleep/sleep/worker/requirements.txt
python-dotenv
flask_sqlalchemy
flask
pika
openai
SQLAlchemy
psycopg2-binary
